Throw mechanic
When holding input "Release" the object that you are grabbing (with the right hand on items and with both on the others objects) the object gets throw with some force.
"items" and "grabbables" get thrown forward and up.
"Draggables" are getting push away from player.
Switch keys on input. "Interact" is "E" key. "Release" is "R". "SwitchCharacters" is "Q"