Item sharing and a bugfix
a bug was discovered that allowed a character to take an item from other character's hand if it was already close.
Now Characters can share items, meaning that a cahracter can grab an item from another character AND update the other character about the exchange. Dog grabs first from human right hand and then from left, not from the closes of those hands.