AxeJump and Jump bugfix
AxeJump is added. When player grabs the axe when it is stuck to a wall, he performs an AxeJump , a kind of double jump.
InteractControl mistake fixed that called SendMessage("OnGrab") when it was released and not when it was grabbed.
Jump bug fixed that allow a jittered flying when pressing jump repeatably on air. Now groundCheck has specific layers.