Changes and Bugfixes
CutDownTree has hitpoints and takes damage, now takes more axe cuts to fall.
Axe Jump requires pressing jump, in a was it is performed with "Interact" and "Jump" inputs.
"InteractCollider" also gets layers.
"Call" input is now key  'c'.
Added 2 layers.  1)"IgnorePlayer" that ignores human and dog. 2)"ObjectSafeArea" that is added to a child gameObject of targeted gameObject with same collider but on trigger. This collider in this layer help on reseting its parent Layer after player hand release.
(As said above) A mechanic was added. When player releases and object (item and grabbable) it is ignored by the player's collider until they are not colliding, preventing from pushing one another. This help a lot with Throwing object but may change if we don't like the change