Checkpoint and many fixes
Checkpoints are added that will reset specific gameObjects and reset player (Need a lot of work).
ControlManager now saves if the behaviour that is trying to enable was enabled before.
InteractColliders removes transforms that are missing.
PlayerMovement and LedgeGrab send multiple raycasts, ensuring that raycast hit the important targets.
Added "PlayerHuman" and "PlayerDog" tags.