holding Interact and Release indicators and bugfix
Holding Interact and Release will show progress with UI.
fix: grabbables and draggables are getting released from both hands.
Bonus: Holding "interact" will do a special interact with the object IF it has a script with "OnHoldInteract" method, else it will perform a normal interact/grab. (This is why MessageArgs class was added, to test if sendmessage was received).