Camera follow and boundaries
Cinemachine virtual camera now has boundaries. These boundaries are controlled by a collider in "LevelSettings" gameobject in Hierarchy.
CameraFollowPivot is added to replace
old script that allowed the object to follow the player. This now follows the player if it's inside the boundaries and not follow on the axis that is not.