ExitSceneCollider and fixes
Added ExitSceneCollider. on player collide the chosen scene start to load.
(This is basically the "goal" object).
| Fixes:
Interact Indicator now doesn't roate when player does.
Player was able to Jump mid air because of its interactCollider's layer.
Stopping Coroutines correclty on InteractControl.
 --|--
Today is the 1 month mark of the project.