DogBehaviour
Dog "barks" at the human if he cant be reached and the recalculates if it can. Bark are describes with "confusion" and "sadness" but there are no special bark applications yet (sad bark etc).
"Call" script tells dog if it wants to follow with a boolean and now by diabling dog's scripts.