Frozen Cameras
"CameraOnBox" script. When the active playable cahracter walk into the trigger collider of the script the camera priority cahnges and camera swaps with a specified still camera. Most of the times it will be a zoomed out camera that shows a large area in one picture. On exit of the area the cmera get back to original.