Swap items on hands and fixes
Swapping items in your hand is availiable with "SwapHand" input.
 "SwapHand" and "Inventory" inputs share the same key and joystick input. This is to compact and have less inputs . Tapping will swap items and holding (with a speccified timer) will open the inventory.
---
Changes on "Inventory" script for the hold button change.
Axe will go to inventory upon release (not throw) if there is an empty slot.
"OnReleaseInventory" is called via sendMessage method and will put the object (after getting released) to the inventory.