Bugfixes
On "Axe" first it checks if an "Inventory" script exist and then checks if is active.
On "Call" it searchs for the dog (was problem on checkpoint reset).
On "InteractControl" checks if is holding item upon getting ready to throw.
On "ControlManager" there is a method (also called by the checkpoint) that recalculates the sets human and dog behaviours for the "new" instances.
On "Checkpoint" players instanciate and reseted correctly ever after multiple resets.