Checkpoint 3
"Checkpoint3" is a new checkpoint system.
It works simular to "Checkpoint" but with some noticable diffferences. Instead of setting up the list of GameObjects that you want to reset in the inspector, Now it resets objects with "Resetable" Component. Upon reaching the checkpoint all of these GameObjects are duplicated and the clone is set to "not original" and inactive. Upon reset, the "originals" are getting destroyed and the "not originals" become "originals" and active, creating a reset illusion.