Adding Resetable to prefabs and adjusting Checkpoint3
The "Resetable" component is added to every prefab that is needed.
Changed "Checkpoint3" on getting all of these components of the scene. With previous code we were giving access not only to the activ and inactive GameObjects of the Hierarchy but it was also searching in the memory a.k.aa the prefabs. this means it was changing the prefabs' Resetable.
This is now changed so it will only search in the Hierarchy.