Trampoline and Locked camera fix
Trampoline prefab was added. It has a physic material with high bounciness.
Fixed a bug: when switching charachers, one is inside the lock camera "CameraOnBox" and one outside, there wasn't a transition and the camera stayed frozen. Now locked camera checks if the active player is inside the boundaries orelse it changes to lower priority.