Rope mechanics are static and as item
2 prefabs are added.
Rope (static/anchored) is 2 units long with 11 joints. The one end is anchored to worldspace and the other end has tag: draggable.
This mean that the human can grab the one tip and drag the rope, or even better grab it to swing or constain himself. Rope has a lenght limit.
Rope (item) is 5 joints long and it has no use for now.