HideInfrontOf script
When this script is attached to a gameObject with a collider/trigger, when a trigger happens with another object with the wanted tags it disables all its renderers. Simple application of this is by putting an object and the trigger farther from the camera ("behind the object"). When the player with tag "PlayerHuman" passes "behind" the object , the object will vanish and when the player leaves from behind it , it reappears.