Test level 2 scene completed nad fixes
Test level 2 scene completed.
Upon enabling "InteractControl" it checks for what the character is holding fixing missing references.
"RestartScene" method reset to checkpoint on default and doesn't reload the scene.
"Checkpoint3" sets itself as the checkpoint ref on LevelSettings upon reaching it.
On "CutDownTree", "Resetable" component is added to the log when thee gets cut down. This prevent log duplicates.