Daughter and inventory
"Girl" script was added.
Human has a daughter that can carry on his back and only on his back. It has a simular behaviour as "Item" when it comes to the inventory.
Inventory's methods were cahnged to take more parametersand allow special interactions. You can force an item into a slot or pick any availiable or even don't at to invetory if slot is full.