Options on main menu and manager prefabs
A modified prefab of the pause menu was added in main menu to change options from there.
"PauseMenuMnager" has a boolean onMainMenu that filters methods that don't wan't to be called. For example, options that change something mid-game. It saves the change but doesn't call the "output".
"MainMenuManager" was changed for the options category and a method that makes transitions through canvas was made more general.
Other small cahnges on previous mentioned scripts.
"GameData" on "SaveManager" are created using the managers instances and not in-class public  references.