Bugfixes
Pause menu will load the player's options on awake. This setups the options for the scene. Previous the options were not implimented until the player opened the menu (quality setting for example).
Audio is not playd when loading data from file.
Fps counter has a refresh rate. This makes it more readable.
Quality dropdown on main menu options was displaying its options upwards and was out of bounds. Now it displays downwards