Input binding testing
"InputManager" script was added.
InputMaster has a new action "Any" with any key as an input. When this action gets performed a combination of the old and the new input system recognizes the key that was pressed and tries to add it as an override button to the binding. This will help for the player to set his custom inputs (on keyboard).