Saving player inputs
Player inputs paths are saved in a 2d matrix [device,action] (on PauseMenuManager).
Then this matrix is saved and loaded from "SaveManager".
"Gamedata" stores player's input paths.
On load, toggles that show the input binded keys are in 2 categories, Keyboard and Controller. They are no in a matrix so it can be displayed in inpector and reference the toggles from hierarchy. These toggle's text gets changed to display the correct binded key (or be empty if there is none).